updates and patches for devsummit

+++ ai-super-reference-kit-2023/pharmaceutical_manufacturing_business/medication_qaqc/src/model.py
@@ -9,7 +9,7 @@ class TrainPayload_IPEX(BaseModel):
     data_folder: str=None
     neg_class: int
     modeldir: str=None
-    learning_rate: int
+    learning_rate: float
     epochs: int
     data_aug: int
